Backend/requirements.txt
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
ecdsa==0.19.1
etelemetry==0.3.1
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.1
fitz==0.0.1.dev2
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pathlib==1.0.1
polars==1.38.1
polars-runtime-32==1.38.1
prov==2.1.1
puremagic==2.0.2
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.6.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
traits==7.1.0
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
